# packages/shared-contracts/openapi/cloudhelm.openapi.yaml
openapi: 3.1.0
info:
  title: CloudHelm Platform API
  description: CloudHelm 平台 API。M2 提供真实数据库驱动的项目、任务、需求、设计、审批和事件底座。
  version: 0.2.0
paths:
...
  /api/tasks/{task_id}/agent-runs:
    post:
      tags:
      - agent-runs
      summary: 创建 AgentRun 记录（M2 内部联调）
      description: 写入 AgentRun 记录；不表示 Agent 已自动执行。
      operationId: create_agent_run_api_tasks__task_id__agent_runs_post
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRunCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRunRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
      - agent-runs
      summary: 分页读取任务 AgentRun
      description: 读取某任务下的 AgentRun 记录。
      operationId: list_agent_runs_api_tasks__task_id__agent_runs_get
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: 分页大小。
          default: 50
          title: Limit
        description: 分页大小。
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: offset cursor。
          title: Cursor
        description: offset cursor。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_AgentRunRead_'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/agent-runs/{run_id}:
    get:
      tags:
      - agent-runs
      summary: 读取 AgentRun
      description: 读取单个 AgentRun。
      operationId: get_agent_run_api_agent_runs__run_id__get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRunRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/tasks/{task_id}/approvals:
    post:
      tags:
      - approvals
      summary: 创建审批请求（M2 内部联调）
      description: 创建 ApprovalRequest；M2 不自动触发工具执行。
      operationId: create_approval_api_tasks__task_id__approvals_post
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalRequestCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalRequestRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/approvals:
    get:
      tags:
      - approvals
      summary: 分页读取审批请求
      description: 分页读取审批请求。
      operationId: list_approvals_api_approvals_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ApprovalStatus'
          - type: 'null'
          description: 按状态过滤。
          title: Status
        description: 按状态过滤。
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: 分页大小。
          default: 50
          title: Limit
        description: 分页大小。
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: offset cursor。
          title: Cursor
        description: offset cursor。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_ApprovalRequestRead_'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/approvals/{approval_id}:
    get:
      tags:
      - approvals
      summary: 读取审批请求
      description: 读取单个 ApprovalRequest。
      operationId: get_approval_api_approvals__approval_id__get
      parameters:
      - name: approval_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Approval Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalRequestRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/approvals/{approval_id}/approve:
    post:
      tags:
      - approvals
      summary: 通过审批请求
      description: 通过审批并写入 ApprovalApproved 事件。
      operationId: approve_approval_api_approvals__approval_id__approve_post
      parameters:
      - name: approval_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Approval Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/DecisionRequest'
              - type: 'null'
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalRequestRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/approvals/{approval_id}/reject:
    post:
      tags:
      - approvals
      summary: 拒绝审批请求
      description: 拒绝审批并写入 ApprovalRejected 事件。
      operationId: reject_approval_api_approvals__approval_id__reject_post
      parameters:
      - name: approval_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Approval Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/DecisionRequest'
              - type: 'null'
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalRequestRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    AgentRunCreate:
      properties:
        agent_type:
          type: string
          minLength: 1
          title: Agent Type
          description: Agent 类型。
        status:
          $ref: '#/components/schemas/AgentRunStatus'
          description: 运行状态。
          default: pending
        model_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Model Name
          description: 模型名称。
        prompt_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Prompt Hash
          description: Prompt 哈希。
        input_tokens:
          type: integer
          minimum: 0.0
          title: Input Tokens
          description: 输入 token 数。
          default: 0
        output_tokens:
          type: integer
          minimum: 0.0
          title: Output Tokens
          description: 输出 token 数。
          default: 0
        cost_usd:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Cost Usd
          description: 估算成本。
          default: '0'
      type: object
      required:
      - agent_type
      title: AgentRunCreate
      description: '开发/内部联调用 AgentRun 创建请求体。


        M2 该接口只写入运行记录，不表示 Agent 已具备自动执行能力。'
    AgentRunRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        task_id:
          type: string
          format: uuid
          title: Task Id
        agent_type:
          type: string
          title: Agent Type
        status:
          $ref: '#/components/schemas/AgentRunStatus'
        model_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Model Name
        prompt_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Prompt Hash
        input_tokens:
          type: integer
          title: Input Tokens
        output_tokens:
          type: integer
          title: Output Tokens
        cost_usd:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Cost Usd
        started_at:
          type: string
          format: date-time
          title: Started At
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
      type: object
      required:
      - id
      - task_id
      - agent_type
      - status
      - model_name
      - prompt_hash
      - input_tokens
      - output_tokens
      - cost_usd
      - started_at
      - finished_at
      title: AgentRunRead
      description: AgentRun 响应结构。
    AgentRunStatus:
      type: string
      enum:
      - pending
      - running
      - succeeded
      - failed
      - cancelled
      title: AgentRunStatus
      description: AgentRun 状态。
    ApprovalRequestCreate:
      properties:
        action:
          type: string
          minLength: 1
          title: Action
          description: 申请执行的动作。
        risk_level:
          $ref: '#/components/schemas/RiskLevel'
          description: 动作风险等级。
        reason:
          type: string
          minLength: 1
          title: Reason
          description: 申请原因。
        requested_by_agent_run_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Requested By Agent Run Id
          description: 发起审批的 AgentRun。
      type: object
      required:
      - action
      - risk_level
      - reason
      title: ApprovalRequestCreate
      description: 开发/内部联调用审批创建请求体。
    ApprovalRequestRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        task_id:
          type: string
          format: uuid
          title: Task Id
        action:
          type: string
          title: Action
        risk_level:
          $ref: '#/components/schemas/RiskLevel'
        reason:
          type: string
          title: Reason
        status:
          $ref: '#/components/schemas/ApprovalStatus'
        requested_by_agent_run_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Requested By Agent Run Id
        decided_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Decided By
        decided_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Decided At
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - task_id
      - action
      - risk_level
      - reason
      - status
      - requested_by_agent_run_id
      - decided_by
      - decided_at
      - created_at
      title: ApprovalRequestRead
      description: 审批请求响应结构。
    ApprovalStatus:
      type: string
      enum:
      - pending
      - approved
      - rejected
      - expired
      title: ApprovalStatus
      description: ApprovalRequest 状态。
    DecisionRequest:
      properties:
        actor_id:
          type: string
          minLength: 1
          title: Actor Id
          description: 操作人或组件标识。
          default: user
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
          description: 操作原因。
      type: object
      title: DecisionRequest
      description: 审批、通过或要求修改类动作的请求体。
    ErrorResponse:
      properties:
        code:
          type: string
          title: Code
          description: 稳定错误码。
        message:
          type: string
          title: Message
          description: 面向调用方的错误描述。
        detail:
          anyOf:
          - {}
          - type: 'null'
          title: Detail
          description: 调试细节或字段错误。
        trace_id:
          type: string
          title: Trace Id
          description: 链路追踪 ID。
      type: object
      required:
      - code
      - message
      - trace_id
      title: ErrorResponse
      description: 统一错误响应结构。
...
    PageInfo:
      properties:
        limit:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: Limit
          description: 本次最多返回条数。
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
          description: 下一页游标；为空表示无更多数据。
      type: object
      required:
      - limit
      title: PageInfo
      description: '分页元数据。


        M2 使用 offset cursor 的最小实现，后续如需高并发滚动查询可升级为

        基于 `created_at,id` 的稳定游标。'
    PageResponse_AgentRunRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AgentRunRead'
          type: array
          title: Items
          description: 当前页数据。
        page:
          $ref: '#/components/schemas/PageInfo'
          description: 分页信息。
      type: object
      required:
      - items
      - page
      title: PageResponse[AgentRunRead]
    PageResponse_ApprovalRequestRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ApprovalRequestRead'
          type: array
          title: Items
          description: 当前页数据。
        page:
          $ref: '#/components/schemas/PageInfo'
          description: 分页信息。
      type: object
      required:
      - items
      - page
      title: PageResponse[ApprovalRequestRead]
...
    RiskLevel:
      type: string
      enum:
      - L0
      - L1
      - L2
      - L3
      - L4
      title: RiskLevel
      description: CloudHelm 工具和任务风险等级。
...
servers:
- url: http://127.0.0.1:18080
  description: 本地开发平台 API